+++ framework/jquery/validators/EmailValidatorJqueryClientScript.php
@@ -1,0 +1,69 @@
+<?php
+
+/**
+ * @link https://www.yiiframework.com/
+ * @copyright Copyright (c) 2008 Yii Software LLC
+ * @license https://www.yiiframework.com/license/
+ */
+
+declare(strict_types=1);
+
+namespace yii\jquery\validators;
+
+use yii\base\Model;
+use yii\helpers\Json;
+use yii\validators\client\ClientValidatorScriptInterface;
+use yii\validators\EmailValidator;
+use yii\validators\PunycodeAsset;
+use yii\validators\ValidationAsset;
+use yii\validators\Validator;
+use yii\web\JsExpression;
+use yii\web\View;
+
+/**
+ * EmailValidatorJqueryClientScript provides client-side validation script generation for email attributes.
+ *
+ * This class implements {@see ClientValidatorScriptInterface} to supply client-side validation options and register the
+ * corresponding JavaScript code for email validation in Yii2 forms using jQuery.
+ *
+ * @template T of EmailValidator
+ * @implements ClientValidatorScriptInterface<T>
+ *
+ * @author Wilmer Arambula <terabytesoftw@gmail.com>
+ * @since 2.2.0
+ */
+class EmailValidatorJqueryClientScript implements ClientValidatorScriptInterface
+{
+    public function getClientOptions(Validator $validator, Model $model, string $attribute): array
+    {
+        $options = [
+            'pattern' => new JsExpression($validator->pattern),
+            'fullPattern' => new JsExpression($validator->fullPattern),
+            'allowName' => $validator->allowName,
+            'message' => $validator->getFormattedClientMessage(
+                $validator->message,
+                ['attribute' => $model->getAttributeLabel($attribute)],
+            ),
+            'enableIDN' => (bool) $validator->enableIDN,
+        ];
+
+        if ($validator->skipOnEmpty) {
+            $options['skipOnEmpty'] = 1;
+        }
+
+        return $options;
+    }
+
+    public function register(Validator $validator, Model $model, string $attribute, View $view): string
+    {
+        ValidationAsset::register($view);
+
+        if ($validator->enableIDN) {
+            PunycodeAsset::register($view);
+        }
+
+        $options = $this->getClientOptions($validator, $model, $attribute);
+
+        return 'yii.validation.email(value, messages, ' . Json::htmlEncode($options) . ');';
+    }
+}